Move a public inner Class to its own compilation unit.
